Wire separation and Timer fixes

+++ TPLTurnout.h
@@ -2,5 +2,5 @@
 class TPLTurnout {
   public :
   static void begin();
-  static short slowSwitch( byte id, bool left, bool expedite);
+  static int slowSwitch( byte id, bool left, bool expedite);
   };
